More changes towards duplicating gates.
