+++ circuit-editor/src/components/ToffoliGate.vue
@@ -171,6 +171,42 @@
           <td class="no-resize-cell"></td>
         </tr>
         <tr>
+          <td colspan="3" width="300px" class="td-2nd-modal">
+            Control Qubit - 'q, s' based <br/>javascript expression:
+          </td>
+          <td colspan="3" width="400px" class="td-2nd-modal">
+            <b-form-input min="0" v-model="controlExpression" placeholder="" type="text" id="ctrl-qbit" style="min-width:400px;"></b-form-input>
+          </td>
+          <td class="no-resize-cell"></td>
+        </tr>
+        <tr>
+          <td colspan="3" width="300px" class="td-2nd-modal">
+            Control State - 'q, s' based <br/>javascript expression:
+          </td>
+          <td colspan="3" width="400px" class="td-2nd-modal">
+            <b-form-input min="0" v-model="controlstateExpression" placeholder="" type="text" id="ctrl-state" style="min-width:400px;"></b-form-input>
+          </td>
+          <td class="no-resize-cell"></td>
+        </tr>
+        <tr>
+          <td colspan="3" width="300px" class="td-2nd-modal">
+            Control-2 Qubit - 'q, s' based <br/>javascript expression:
+          </td>
+          <td colspan="3" width="400px" class="td-2nd-modal">
+            <b-form-input min="0" v-model="control2Expression" placeholder="" type="text" id="ctrl-2-qbit" style="min-width:400px;"></b-form-input>
+          </td>
+          <td class="no-resize-cell"></td>
+        </tr>
+        <tr>
+          <td colspan="3" width="300px" class="td-2nd-modal">
+            Control-2 State - 'q, s' based <br/>javascript expression:
+          </td>
+          <td colspan="3" width="400px" class="td-2nd-modal">
+            <b-form-input min="0" v-model="controlstate2Expression" placeholder="" type="text" id="ctrl-2-state" style="min-width:400px;"></b-form-input>
+          </td>
+          <td class="no-resize-cell"></td>
+        </tr>
+        <tr>
           <td colspan="6" class="td-2nd-modal">
           </td>
           <td class="no-resize-cell">
@@ -203,6 +239,10 @@ export default {
     return {
       controlNew2: this.control2,
       controlstateNew2: this.controlstate2,
+      controlExpression: this.control,
+      controlstateExpression: this.controlstate,
+      control2Expression: this.control2,
+      controlstate2Expression: this.controlstate2,
       options: [
         { value: '1', text: '|1⟩' },
         { value: '0', text: '|0⟩' },
@@ -294,6 +334,10 @@ export default {
         'qbitLast': this.qbitLast,
         'qbitConditionExpression': this.qbitConditionExpression,
         'conjugateConditionExpression': this.conjugateConditionExpression,
+        'controlExpression': this.controlExpression,
+        'controlstateExpression': this.controlstateExpression,
+        'control2Expression': this.control2Expression,
+        'controlstate2Expression': this.controlstate2Expression,
       });
       promise.then(
         // eslint-disable-next-line no-unused-vars
